requirements.txt
pyyaml
feedparser
python-dateutil
requests
zulip